backlog: New structure for storing items
The previous version would make a separate request for every item in the
backlog. If the backlog had 10 episodes from a show, it would make 10
requests to Trakt servers to clear the items. Since this is highly
inefficient, and might be subject to rate limiting, the new format will
allow for backlog clearing with a single API request.

Closes #96