backlog: Add UnknownItems class
If we are unable to find the correct trakt id for a movie or a show
when trying to clear the backlog, it will fail to be added to history.
Earlier, it was just kept around for 30 days, and then deleted. This
commit introduces the class 'UnknownItems' to keep track of such media.
Now, invalid media will be removed from backlog, and added to this class
so that user may manually specify what to do with each item. For example
they may choose to simply delete the item, and with it, our knowledge of
the watch timestamp. Or, they may specify a Trakt ID for the item, so as
to not lose the played history for that media.